Improve change detection logic.
Notify the base class whenever the list of items might have changed. Previously
we did not notfiy for the case where the list of items was deleted.

Fixes an NPE on Nexus 9 running 5.0.1. Could in theory have caused problems
on any device, but for some reason simulated devices don't exhibit the bug.